refactor: use consistent directory structure
