+++ .gitignore
@@ -1,7 +1,6 @@
 .nyc_output
 .vscode
 build
-out
 node_modules
 package-lock.json
 .coverage
