+++ .jsdoc.js
@@ -31,7 +31,7 @@ module.exports = {
   source: {
     excludePattern: '(^|\\/|\\\\)[._]',
     include: [
-      'out/src'
+      'build/src'
     ],
     includePattern: '\\.js$'
   },
