+++ synth.py
@@ -6,5 +6,5 @@ logging.basicConfig(level=logging.DEBUG)
 AUTOSYNTH_MULTIPLE_COMMITS = True
 
 common_templates = gcp.CommonTemplates()
-templates = common_templates.node_library(source_location='out/src')
+templates = common_templates.node_library(source_location='build/src')
 s.copy(templates, excludes=['README.md'])
